+++ figureo/standard/text.py
@@ -7,10 +7,10 @@
 # be prosecuted under federal law. Its content is company confidential.
 # =============================================================================
 
-import configo
+import configos
 import iamraw
 import pdfminer.layout
-import utila
+import utilo
 
 # TODO: REMOVE HORIZONTAL AND VERTICAL LINES TO AVOID DETECTING TABLES AS
 # FIGURE?
@@ -50,7 +50,7 @@ def text_figures(
     ]
     # TODO: INVESTIGATE LATER
     # skip very small items
-    result = [item for item in result if utila.rect_width(item.bounding) > 25.0]
+    result = [item for item in result if utilo.rect_width(item.bounding) > 25.0]
     second_look = figures_missing(items, result)
     if second_look:
         result.extend(second_look)
@@ -61,18 +61,18 @@ def figures_missing(items, done) -> list:
     """Backup strategy to detect all LTFigures which are not part of
     extracted text figures."""
     done = [item.bounding for item in done]
-    figures = utila.select_type(items, selector=pdfminer.layout.LTFigure)
+    figures = utilo.select_type(items, selector=pdfminer.layout.LTFigure)
     result = []
     for item in figures:
-        if utila.rectangles_intersecting(done, item.bbox):
+        if utilo.rectangles_intersecting(done, item.bbox):
             # LTFigure is already part of a text figure
             continue
         matching = [
             test.bbox
             for test in items
-            if utila.rect_intersecting(item.bbox, test.bbox)
+            if utilo.rect_intersecting(item.bbox, test.bbox)
         ]
-        bounding = utila.rect_max([item.bbox] + matching)
+        bounding = utilo.rect_max([item.bbox] + matching)
         figure = iamraw.Figure(bounding=bounding)
         result.append(figure)
     return result
@@ -83,16 +83,16 @@ def bounding_valid(bounding: tuple, items, width_min, height_min, area_min) -> b
     positive renderings."""
     if not textonly(bounding, items):
         return True
-    if utila.rect_size(bounding) < area_min:
+    if utilo.rect_size(bounding) < area_min:
         return False
-    if utila.rect_width(bounding) >= width_min:
+    if utilo.rect_width(bounding) >= width_min:
         return True
-    if utila.rect_height(bounding) >= height_min:
+    if utilo.rect_height(bounding) >= height_min:
         return True
     return False
 
 
-CONTENT_INVALID_RATE_MAX = configo.HolyTable(items=[
+CONTENT_INVALID_RATE_MAX = configos.HolyTable(items=[
     (5, 0.1),
     (10, 0.2),
     (15, 0.2),
@@ -107,11 +107,11 @@ def content_valid(bounding, content, invalids) -> bool:
         return True
     content = [
         item for item in content
-        if utila.rect_intersecting(bounding, item.bbox)
+        if utilo.rect_intersecting(bounding, item.bbox)
     ]
     invalids = [
         item for item in invalids
-        if utila.rect_intersecting(bounding, item.bbox)
+        if utilo.rect_intersecting(bounding, item.bbox)
     ]
     rate = len(invalids) / len(content)
     if rate > CONTENT_INVALID_RATE_MAX(len(content)):
@@ -124,7 +124,7 @@ def splitby_breaker(items, breaker):
     if not breaker:
         return [items]
     breaker = sorted(breaker)
-    grouped = utila.Buckets(border=breaker)
+    grouped = utilo.Buckets(border=breaker)
     grouped.selector = lambda x: (x.bbox[1] + x.bbox[3]) / 2
     for item in items:
         grouped.add(item)
@@ -138,7 +138,7 @@ def textonly(bounding, items: list) -> bool:
         item for item in items
         if not isinstance(item, pdfminer.layout.LTTextBoxHorizontal)
     ]
-    if any((utila.rect_inside(
+    if any((utilo.rect_inside(
             bounding,
             item.bbox,
             diff=10,
@@ -147,23 +147,23 @@ def textonly(bounding, items: list) -> bool:
     return True
 
 
-CLUSTER_SIZE_MIN = configo.HV_INT_PLUS(default=25)
+CLUSTER_SIZE_MIN = configos.HV_INT_PLUS(default=25)
 
 
 def determine_clusters(
     items: list,
     cluster_size_min=CLUSTER_SIZE_MIN,
 ):
-    bucket = utila.Buckets(utila.ranges(0, 1000, 15), sorting=True)
+    bucket = utilo.Buckets(utilo.ranges(0, 1000, 15), sorting=True)
     for item in items:
         start, end = item.bbox[1], item.bbox[3]
         # left to right to ensure that line is marked more than one
         # keep y-expansion in calculation
-        # for _ in utila.ranges(item.bbox[0], item.bbox[2], step=50):
-        for coordinate in utila.ranges(start, end, step=5):
-            bucket.add(utila.roundme(coordinate))
+        # for _ in utilo.ranges(item.bbox[0], item.bbox[2], step=50):
+        for coordinate in utilo.ranges(start, end, step=5):
+            bucket.add(utilo.roundme(coordinate))
     # select areas with enough items
-    content = utila.groupby_neighbors(bucket)
+    content = utilo.groupby_neighbors(bucket)
     # TODO: CHECK CLUSTER_SIZE_MIN CAUSE SET REMOVES ITEMS OUT OF CONTENT
     selected = [set(item) for item in content if len(item) >= cluster_size_min]
     # prepare result
@@ -175,9 +175,9 @@ def determine_clusters(
 
 
 # A normal line of text is between 10-13 height
-AREA_START_HEIGHT_MAX = configo.HV_FLOAT_PLUS(default=25.0)
+AREA_START_HEIGHT_MAX = configos.HV_FLOAT_PLUS(default=25.0)
 # Covering less than this is detected as text line start
-AREA_COVERING_RATE_MIN = configo.HV_PERCENT_PLUS(default=35.0)
+AREA_COVERING_RATE_MIN = configos.HV_PERCENT_PLUS(default=35.0)
 
 
 def determine_cluster_rectangle(
@@ -198,7 +198,7 @@ def determine_cluster_rectangle(
     # determine text start
     start_area = (x0, y0, x1, y0 + first)
     firstline = [
-        item for item in items if utila.rect_inside(start_area, item.bbox)
+        item for item in items if utilo.rect_inside(start_area, item.bbox)
     ]
     textline_in_figure = hanging_textline(start_area, firstline)
     if textline_in_figure:
@@ -206,18 +206,18 @@ def determine_cluster_rectangle(
         for item in firstline:
             incluster.remove(item)
     # determine figure bounding
-    result = utila.rect_max([item.bbox for item in incluster])
+    result = utilo.rect_max([item.bbox for item in incluster])
     return result
 
 
 def hanging_textline(area, firstline) -> bool:
-    buckets = utila.Buckets(border=utila.ranges(
+    buckets = utilo.Buckets(border=utilo.ranges(
         start=area[0],
         stop=area[2],
         step=20.0,
     ))
     for item in firstline:
-        for x in utila.ranges(start=item.bbox[0], stop=item.bbox[2], step=5.0):
+        for x in utilo.ranges(start=item.bbox[0], stop=item.bbox[2], step=5.0):
             buckets.add(x)
     # 1/3 2/3
     start, rest = splitby_percent(buckets, percent=0.35)