+++ figureo/utils.py
@@ -10,7 +10,7 @@
 import io
 
 import PIL.Image
-import utila
+import utilo
 
 DumpedFigureInformation = list[tuple[str, bytes]]
 
@@ -31,16 +31,16 @@ IMAGE_HEIGHT_MAX = 768
 
 
 def rawfigure_frombounding(bbox, mode=RGBA, background=WHITE) -> PIL.Image:
-    width = utila.rect_width(bbox)
-    height = utila.rect_height(bbox)
+    width = utilo.rect_width(bbox)
+    height = utilo.rect_height(bbox)
     # limit max figure size to avoid too much memory consumption
-    width = utila.mins(width, IMAGE_WIDTH_MAX)
-    height = utila.mins(height, IMAGE_HEIGHT_MAX)
+    width = utilo.mins(width, IMAGE_WIDTH_MAX)
+    height = utilo.mins(height, IMAGE_HEIGHT_MAX)
     # ensure positive figure size
     if width < 0 or height < 0:
-        utila.error(f'negative figure size: {width} {height}')
-    width = utila.maxs(width, 1)
-    height = utila.maxs(height, 1)
+        utilo.error(f'negative figure size: {width} {height}')
+    width = utilo.maxs(width, 1)
+    height = utilo.maxs(height, 1)
     # figure image size
     size = (int(width), int(height))
     raw = PIL.Image.new(mode, size, color=background)