+++ tests/test_cleanup.py
@@ -9,27 +9,27 @@
 
 import os
 
-import power
+import hoverpower
 import serializeraw
-import utila
-import utilatest
+import utilo
+import utilotest
 
 import tests
 import tests.test_render
 
 
-@utilatest.requires(power.BACHELOR051_PDF)
+@utilotest.requires(hoverpower.BACHELOR051_PDF)
 def test_bachelor51page29_cleanup(td, mp):
     generated = os.path.join(
-        power.link(power.BACHELOR051_PDF),
+        hoverpower.link(hoverpower.BACHELOR051_PDF),
         'rawmaker__images_images',
     )
     tests.test_render.run_standard(
-        power.BACHELOR051_PDF,
+        hoverpower.BACHELOR051_PDF,
         pages=29,
         mp=mp,
     )
-    utila.copy_content(
+    utilo.copy_content(
         generated,
         td.tmpdir.join('rawmaker__images_images'),
         unlock=True,
@@ -41,7 +41,7 @@ def test_bachelor51page29_cleanup(td, mp):
     )
     images = serializeraw.load_image_infos_frompath('rawmaker__images_images')
     # ensure that some images are hidden by extract figure
-    assert any(item.hidden for item in utila.flatten_content(images))
+    assert any(item.hidden for item in utilo.flatten_content(images))
 
 
 def run_cleanup(path, pages, mp) -> list:
@@ -53,15 +53,15 @@ def run_cleanup(path, pages, mp) -> list:
     return images
 
 
-@utilatest.longrun
-@utilatest.requires(power.BACHELOR051_PDF)
+@utilotest.longrun
+@utilotest.requires(hoverpower.BACHELOR051_PDF)
 def test_bachelor51p30_hide_images(td, mp):
-    pdf = power.BACHELOR051_PDF
-    utila.run(f'rawmaker -i {pdf} --images --pages=30 -o {td.tmpdir}')
+    pdf = hoverpower.BACHELOR051_PDF
+    utilo.run(f'rawmaker -i {pdf} --images --pages=30 -o {td.tmpdir}')
     workdir = td.tmpdir.join('rawmaker__images_images')
     cmd = f'-i {pdf} -i {workdir} -i {td.tmpdir} -o {td.tmpdir} --pages=30'
     tests.run(cmd, mp=mp)
     images = serializeraw.load_image_infos_frompath(workdir)
-    images = utila.flatten_content(images)
+    images = utilo.flatten_content(images)
     hidden = [item for item in images if item.hidden]
     assert len(hidden) == 4