+++ tests/test_nofigure.py
@@ -9,23 +9,24 @@
 
 import os
 
-import power
+import hoverpower
 import pytest
-import utilatest
+import utilotest
 
 import tests
 
 
 @pytest.mark.usefixtures('td')
 @pytest.mark.parametrize('source, pages', [
-    pytest.param(power.DISS266_PDF, '156,168,204', id='small_text_elements'),
-    pytest.param(power.DISS266_PDF, '27,28,61', id='diss266'),
-    pytest.param(power.HOME050_PDF, '31', id='home50p31_formula'),
-    pytest.param(power.MASTER091B_PDF, '19', id='master091bp19'),
+    pytest.param(
+        hoverpower.DISS266_PDF, '156,168,204', id='small_text_elements'),
+    pytest.param(hoverpower.DISS266_PDF, '27,28,61', id='diss266'),
+    pytest.param(hoverpower.HOME050_PDF, '31', id='home50p31_formula'),
+    pytest.param(hoverpower.MASTER091B_PDF, '19', id='master091bp19'),
 ])
 def test_nofigure(source, pages, mp):
-    utilatest.fixture_requires(source)
-    generated = power.link(source)
+    utilotest.fixture_requires(source)
+    generated = hoverpower.link(source)
     tests.run(
         f'-i {generated} -i {source} --standard --pages={pages}',
         mp=mp,