+++ tests/test_render.py
@@ -9,11 +9,11 @@
 
 import os
 
-import power
+import hoverpower
 import pytest
 import serializeraw
-import utila
-import utilatest
+import utilo
+import utilotest
 
 import tests
 
@@ -23,10 +23,10 @@ RENDER_ACCURACY = 5.0
 def test_figures_run_bachelor56page27(td, mp):
     """Ensure that text below, left and right from figure is included
     into figure."""
-    run_standard(power.BACHELOR056_PDF, pages=27, mp=mp)
+    run_standard(hoverpower.BACHELOR056_PDF, pages=27, mp=mp)
     images = serializeraw.load_image_infos_frompath(
         td.tmpdir.join('rawmaker__images_images'))
-    images = utila.flatten_content(images)
+    images = utilo.flatten_content(images)
     assert len(images) == 1
     image = images[0]
     assert image.width >= 217, image.width
@@ -34,10 +34,10 @@ def test_figures_run_bachelor56page27(td, mp):
 
 
 def test_figures_run_bachelor56page19(td, mp):
-    run_standard(power.BACHELOR056_PDF, pages=18, mp=mp)
+    run_standard(hoverpower.BACHELOR056_PDF, pages=18, mp=mp)
     images = serializeraw.load_image_infos_frompath(
         td.tmpdir.join('rawmaker__images_images'))
-    images = utila.flatten_content(images)
+    images = utilo.flatten_content(images)
     assert len(images) == 1
     image = images[0]
     assert image.width >= 461, image.width
@@ -47,7 +47,7 @@ def test_figures_run_bachelor56page19(td, mp):
 @pytest.mark.usefixtures('td')
 def test_figures_skip_dots(mp):
     images = run_standard(
-        power.BACHELOR090_PDF,
+        hoverpower.BACHELOR090_PDF,
         pages='81,82',
         mp=mp,
     )
@@ -57,7 +57,7 @@ def test_figures_skip_dots(mp):
 
 def test_figures_skip_xml(td, mp):  # pylint:disable=W0613
     images = run_standard(
-        power.BACHELOR067_PDF,
+        hoverpower.BACHELOR067_PDF,
         pages='61',
         mp=mp,
     )
@@ -66,11 +66,11 @@ def test_figures_skip_xml(td, mp):  # pylint:disable=W0613
 
 
 @pytest.mark.usefixtures('td')
-@utilatest.requires(power.BACHELOR090_PDF)
+@utilotest.requires(hoverpower.BACHELOR090_PDF)
 def test_figures_double_image(mp):
     """This is an image, not a figure. We have to skip this."""
     images = run_standard(
-        power.BACHELOR090_PDF,
+        hoverpower.BACHELOR090_PDF,
         pages=80,
         mp=mp,
     )
@@ -81,7 +81,7 @@ def test_figures_double_image(mp):
 @pytest.mark.usefixtures('td')
 def test_reg_figure_text_in_figure(mp):
     """Do not include `Diplomarbeit` inside title page figure."""
-    images = run_standard(power.MASTER078_PDF, pages=0, mp=mp)
+    images = run_standard(hoverpower.MASTER078_PDF, pages=0, mp=mp)
     image = images[0]
     assert image.height < 140.0, f'Diplomarbeit included: {image.height}'
 
@@ -89,10 +89,10 @@ def test_reg_figure_text_in_figure(mp):
 @pytest.mark.xfail(reason='software integration')
 @pytest.mark.usefixtures('td')
 def test_master31page4(mp):
-    images = run_standard(power.MASTER031_PDF, pages=4, mp=mp)
+    images = run_standard(hoverpower.MASTER031_PDF, pages=4, mp=mp)
     selected = images[0].bounding
     expected = (297.36, 66.0, 532.92, 257.28)
-    assert utila.nears(selected, expected)
+    assert utilo.nears(selected, expected)
 
 
 @pytest.mark.usefixtures('td')
@@ -102,45 +102,45 @@ def test_master31page10(mp):
     Two Asian characters are handled by rawmaker --image.
     """
     images = run_standard(
-        power.MASTER031_PDF,
+        hoverpower.MASTER031_PDF,
         pages=10,
         mp=mp,
     )
     assert len(images) == 1  # pylint:disable=C2001
     selected = images[0].bounding
     expected = (71.61, 415.7, 524.5, 556.19)
-    assert utila.nears(selected, expected, diff=4.0)
+    assert utilo.nears(selected, expected, diff=4.0)
 
 
 @pytest.mark.xfail(reason='new software')
-@utilatest.longrun
+@utilotest.longrun
 @pytest.mark.usefixtures('td')
 def test_master75page1718(mp):
     images = run_standard(
-        power.MASTER075_PDF,
+        hoverpower.MASTER075_PDF,
         pages='17,18',
         mp=mp,
     )
     # page17
     bounding = images[0].bounding
     expected = (70.85, 135.41, 452.0, 717.15)
-    assert utila.nears(bounding, expected)
+    assert utilo.nears(bounding, expected)
     # page18
     bounding = images[1].bounding
     expected = (70.85, 73.3, 416.39, 725.25)
-    assert utila.nears(bounding, expected)
+    assert utilo.nears(bounding, expected)
 
 
 @pytest.mark.usefixtures('td')
 def test_bachelor51page29_nocleanup(mp):
     images = run_standard(
-        power.BACHELOR051_PDF,
+        hoverpower.BACHELOR051_PDF,
         pages=29,
         mp=mp,
     )
     selected = images[0].bounding
     expected = (86.16, 56.28, 440.16, 281.28)
-    assert utila.nears(selected, expected, diff=RENDER_ACCURACY)
+    assert utilo.nears(selected, expected, diff=RENDER_ACCURACY)
 
 
 @pytest.mark.xfail(reason='new software')
@@ -149,51 +149,51 @@ def test_bachelor37page18(mp):
     """Do not increase valid_area tolerance too much to avoid including
     parts of header and footer into extracted figures."""
     images = run_standard(
-        power.BACHELOR037_PDF,
+        hoverpower.BACHELOR037_PDF,
         pages=18,
         mp=mp,
     )
     selected = images[0].bounding
     expected = (95.03, 63.46, 514.56, 210.87)
-    assert utila.nears(selected, expected)
+    assert utilo.nears(selected, expected)
 
 
 @pytest.mark.xfail(reason='new software')
 @pytest.mark.usefixtures('td')
 def test_bachelor37page23(mp):
     images = run_standard(
-        power.BACHELOR037_PDF,
+        hoverpower.BACHELOR037_PDF,
         pages=23,
         mp=mp,
     )
     boundings = sorted([image.bounding for image in images])
     expected = [(65.52, 559.2, 529.92, 632.27), (70.92, 253.19, 510.36, 383.16)]
-    assert utila.nears(boundings[0], expected[0])
-    assert utila.nears(boundings[1], expected[1])
+    assert utilo.nears(boundings[0], expected[0])
+    assert utilo.nears(boundings[1], expected[1])
 
 
 @pytest.mark.usefixtures('td')
 def test_diss205p102(mp):
     images = run_standard(
-        power.DISS205_PDF,
+        hoverpower.DISS205_PDF,
         pages=102,
         mp=mp,
     )
     selected = images[0].bounding
     expected = (179.13, 250.27, 406.71, 437.01)
-    assert utila.nears(selected, expected)
+    assert utilo.nears(selected, expected)
 
 
 @pytest.mark.usefixtures('td')
 def test_diss205p141(mp):
     images = run_standard(
-        power.DISS205_PDF,
+        hoverpower.DISS205_PDF,
         pages=141,
         mp=mp,
     )
     selected = images[0].bounding
     expected = (81.23, 309.1, 505.51, 450.41)
-    assert utila.nears(selected, expected)
+    assert utilo.nears(selected, expected)
 
 
 # yapf:disable
@@ -213,19 +213,19 @@ def test_bachelor67pagex(page, expected, td, mp):  # pylint:disable=W0613
     page15: Detect single text figures which are build out of a single LTFigure.
     page48: Shrink figure size which is badly printed by pdfprinter.
     """
-    verify(power.BACHELOR067_PDF, page, expected, mp)
+    verify(hoverpower.BACHELOR067_PDF, page, expected, mp)
 
 
 @pytest.mark.usefixtures('td')
 def test_master063p25(mp):
     images = run_standard(
-        power.MASTER063_PDF,
+        hoverpower.MASTER063_PDF,
         pages=25,
         mp=mp,
     )
     selected = images[0].bounding
     expected = (70.85, 277.83, 408.35, 451.08)
-    assert utila.nears(selected, expected, diff=10)
+    assert utilo.nears(selected, expected, diff=10)
 
 
 def verify(source, page, expected, mp):
@@ -236,9 +236,9 @@ def verify(source, page, expected, mp):
     )
     if isinstance(expected, tuple):
         expected = [expected]
-    boxs = utila.sort_leftright_topdown([figure.bounding for figure in images])
+    boxs = utilo.sort_leftright_topdown([figure.bounding for figure in images])
     for bbox, expect in zip(boxs, expected):
-        assert utila.nears(
+        assert utilo.nears(
             bbox,
             expect,
             diff=5.0,
@@ -246,14 +246,14 @@ def verify(source, page, expected, mp):
 
 
 def run_standard(source, pages, mp) -> list:
-    utilatest.fixture_requires(source)
+    utilotest.fixture_requires(source)
     cmd = f'-i {source} --pages={pages} --standard'
-    source = power.link(source)
+    source = hoverpower.link(source)
     if os.path.exists(source):
         cmd = f'{cmd} -i {source}'
     tests.run(cmd, mp=mp)
     if not os.path.exists('rawmaker__images_images'):
         return []
     images = serializeraw.load_image_infos_frompath('rawmaker__images_images')
-    result = utila.flatten_content(images)
+    result = utilo.flatten_content(images)
     return result